Misc cleanup and disable villager move speed adjustment.
